Refactoring in DB plugin
